refactor: extract TransactionsView from Dashboard
